Fix taskkill

+++ Workerman/Process.php
@@ -16,7 +16,7 @@ class Process {
     public string $cmd;
     public array $descriptorspec = array(
         0 => array("pipe", "r"),  // stdin - канал, из которого дочерний процесс будет читать
-        1 => array("pipe", "w"),  // stdout - канал, в который дочерний процесс будет записывать
+        //1 => array("pipe", "w"),  // stdout - канал, в который дочерний процесс будет записывать
         //2 => array("file", "error-output.txt", "a+") // stderr - файл для записи
     );
     public ?array $pipes;
@@ -26,7 +26,7 @@ class Process {
     public array $env = array('some_option' => 'aeiou');
 
     public string $output;
-    public string $result = '';
+    public $result;
 
     public function __construct(string $cmd, ?string $key = null, ?array $descriptorspec = null, ?string $cwd = null, ?array $env = null, ?int $terminate_after = null) {
         $this->pkey = $key;
@@ -48,7 +48,7 @@ class Process {
 
             $write = '<?php print_r($_ENV); ?>';
             //fwrite($this->pipes[0], $write);
-            $this->output = stream_get_contents($this->pipes[1]);
+            //$this->output = stream_get_contents($this->pipes[1]);
             //debug($this->output);
 
             $this->pstatus = proc_get_status($this->process);
@@ -82,8 +82,10 @@ class Process {
     
     public static function killProc($pkey) {
         if ($process = self::getProcess($pkey)) {
-            debug($process);
+            debug(self::isRun($process->pid));
             $process->kill();
+            debug($process);
+            debug(self::isRun($process->pid));
             debug("команда вернула $process->result");
         }
     }
@@ -97,10 +99,41 @@ class Process {
         return $_SESSION['process'] ?? [];
     }
 
+    public static function getTaskList(): array {
+        exec("tasklist /fo list", $out);
+        debug($out);
+        return $_SESSION['process'] ?? [];
+    }
+
     public static function clean() {
         if (isset($_SESSION['process'])) {
             unset($_SESSION['process']);
         }
+    }
+
+    /** tasklist [/s <computer> [/u [<domain>\]<username> [/p <password>]]] [{/m <module> | /svc | /v}] [/fo {table | list | csv}] [/nh] [/fi <filter> [/fi <filter> [ ... ]]]
+    * /fo {table | list | csv}	Указывает формат, используемый для выходных данных. Допустимые значения: Table, List и CSV. Формат выходных данных по умолчанию — Table.
+    * /Fi <filter>	Указывает типы процессов, включаемых в запрос или исключаемых из него. Можно использовать более одного фильтра или использовать подстановочный знак ( \ ) для указания всех задач или имен изображений. Допустимые фильтры перечислены в разделе имена фильтров, операторы и значения этой статьи.
+    */
+            
+    /**
+    * taskkill [/s <computer> [/u [<domain>\]<username> [/p [<password>]]]] {[/fi <filter>] [...] [/pid <processID> | /im <imagename>]} [/f] [/t]
+    * /f	Указывает, что процессы принудительно завершены. Этот параметр не учитывается для удаленных процессов; все удаленные процессы принудительно завершены.
+    * /t	Завершает указанный процесс и все дочерние процессы, запущенные этим процессом.
+    * /PID <processID>	Указывает идентификатор процесса для завершения процесса.
+    */
+    public static function isRun(int $pid): bool {
+        if (stripos(php_uname('s'), 'win') > -1) {
+            exec("tasklist /fo list /fi \"pid eq $pid\"", $out);
+            debug($out);
+            if (count($out) > 1) {
+                return true;
+            }
+        }
+        elseif (posix_kill(intval($pid), 0)) {
+            return true;
+        }
+        return false;
     }
 
     /**
@@ -109,18 +142,25 @@ class Process {
     * It works only under windows, you need a different kill routine on linux.
     * he script terminates the (else endless running) ping process after approximatly 5 seconds.
     */
-    public function kill(): string {
-        $return_value = proc_terminate($this->process);
-        //$return_value = stripos(php_uname('s'), 'win') > -1 ? exec("taskkill /F /T /PID $this->pid") : exec("kill -9 $this->pid"); // вместо proc_terminate($this->process);
-        // Важно закрывать все каналы перед вызовом proc_close во избежание мёртвой блокировки
-        foreach ($this->pipes as $pipe) {
-            fclose($pipe);
-        }
-        $return_value2 = proc_close($this->process);
+    public function kill() {
         if (isset($_SESSION['process']) && isset($_SESSION['process'][$this->getPid()])) {
             unset($_SESSION['process'][$this->getPid()]);
         }
-        return $this->result = "$return_value, $return_value2";
+
+        if (is_resource($this->process)) {
+            $return_value = proc_terminate($this->process);
+            // Важно закрывать все каналы перед вызовом proc_close во избежание мёртвой блокировки
+            foreach ($this->pipes as $pipe) {
+                if (is_resource($pipe)) fclose($pipe);
+            }
+            $return_value2 = proc_close($this->process);
+            return $this->result = "$return_value, $return_value2";
+        }
+
+        // вместо proc_terminate($this->process);
+        stripos(php_uname('s'), 'win') > -1 ? exec("taskkill /f /t /pid $pid", $this->result) : exec("kill -9 $pid", $this->result);
+        return $this->result;
+        //posix_kill($this->pid, SIGKILL);
     }
     
     public function getPid() {
